Refactor ODB2 loading into classes that can be inherited at different sites
