+++ src/CSET/cset_workflow/app/metplus_prep_obs/bin/odb2/__init__.py
@@ -1,0 +1,1 @@
+from .odb2 import PrepODB2Pattern, valid_times_iterator
